# Conflicts:
#	src/main/java/config/ControllerConfig.java
#	src/main/java/config/MvcConfig.java
#	src/main/java/controller/MemberController.java
#	src/main/java/controller/ProductController.java
#	src/main/java/model/member/BuyerDAO.java
#	src/main/java/model/member/BuyerDO.java
#	src/main/java/model/member/SellerDAO.java
#	src/main/java/model/member/SellerDO.java
#	src/main/webapp/WEB-INF/views/header.jsp
#	src/main/webapp/WEB-INF/views/signup.jsp
#	src/main/webapp/WEB-INF/views/signupBuyer.jsp
#	src/main/webapp/WEB-INF/views/signupSeller.jsp
#	src/main/webapp/WEB-INF/web.xml
#	src/main/webapp/css/signupStyle.css

+++ src/main/java/model/member/SellerDO.java
@@ -1,3 +1,4 @@
+//package main.java.model.member;
 package model.member;
 
 public class SellerDO {
@@ -8,7 +9,7 @@ public class SellerDO {
     private String nickname;
     private String passwd;
     private int point;
-    private int tel;
+    private long tel;
     private String regdate;
     private String sellerImg;
     private String address;
@@ -61,11 +62,11 @@ public class SellerDO {
         this.point = point;
     }
 
-    public int getTel() {
+    public long getTel() {
         return tel;
     }
 
-    public void setTel(int tel) {
+    public void setTel(long tel) {
         this.tel = tel;
     }
